Add new DBApi, based on sessions and JSONObjects
Using this setup is is not necessary anymore to use thread locals. The
API becomes a bit cleaner and more readable.

Iterators are used instead of lists.

The records in this API are all from the type JSONObjects, which might
not be the best in terms of performance, because a lot of transformation
has to be done (CHECK!).