Add servlet filter to better react on session timeouts
When a user is logged out with keycloak and a user is doing an API call
then this results in a redirect of the API call to the keycloak login page.
This page is returned to the frontend and results in an error since JSON is
expected. With the filter we want to send a correct error code, which can be
handled by the frontend correctly (http-interceptor).