Fix parentId assignment
The backenderesponse to the getChildren call does not necessarily
only returndocs with the requested parent. When parent is null (root),
the documents of the users assigned groups are also returned, as
they are displayed at the root of the tree. So we have to set their parent
explicitly (to root/null).