Test: fix some failing test The test in ticket (#3047) deleted the  the group 'gruppe_mit_ortsrechten', however This group is used by many tests, so instead
I modified the test to create new document, new group,
assign the new document to the group, assign the new document to user meta2

Then login with meta2 delete the group.
There are already methods for all that,
However the easiest and proper solution is add new group with a document in initial database
so we can delete it directly.