add check for dialog box being closed properly
it has happened that the dialog box asking if changes are supposed to be discarded popped up again after closing it, so that changes had to be discarded twice. This did not provoke any tests to fail because they were not testing for the closed dialog box to remain closed. This is why it is now checked that dialog box doesn't pop up again.