Only use parent in DocumentWrapper
This  avoids the redundant information about the parent. Here we only want to
store the parent information inside the wrapper. Although it seems a good idea
that each document version knows its parent at their time, it leads to confusion
when for example a document has been moved and then reverted, where it would be
moved back to the parent of the published version.