Fix reloading of dataset when using browser back button
In that case it's possible to jump back to an opened document directly instead via the menu button,
where we first are redirected to the form-page and then another redirect to the last opened document.
The fix makes sure that the document is reloaded on a revisit, so that it's also correctly initialized.