Fix getting root nodes from tree
When a tree node has a parent which a user has no read access to, then it's a root node whose parent is not null.
So we cannot use this information to get all root nodes.
We introduce a property isRoot to determine if a node is a root node.