Improve form-guard to prevent loading document twice
When clicking a document in the tree, there happened two events. The normal select event which does load the document and a reload event which
came from the redirect-form.guard.ts. The latter one should only happen when we came from a different page to make sure the form is correctly initialized.