Fix preview image loading when loading another document
When another document is loaded then the form is resetted which leads to removing each preview image after another.
Due to the subscription the form value, we already handle the change although the context is already switched to the other document which we change to.
With debounceTime we give some time before handling the change event.