Add Zabbix report to the UVP profile
Added new files for the Zabbix report behavior, the report component, and the report module for the user-visible-profile (UVP). This is to display the Zabbix report plug-in on the UVP, including HTML and CSS files for the user interface of the report. The profile file was updated to register the Zabbix report behaviour. This update was necessary to allow the management of Zabbix reports within the UVP.