Add Zabbix report endpoint and components
A new feature has been implemented to include Zabbix reporting in our application. The ZabbixReport api has been added to fetch reports related to problems from a given catalog. The corresponding `getProblems` function has been added to the Zabbix service. The response from the api is mapped to the ProblemReportItem class and returned.

On the frontend, `getZabbixReport` function has been added to the UvpResearchService to make requests to the backend. The ZabbixReportComponent has been updated, and it now uses Zabbix reports from the backend.

Test cases are also included to verify the correct behaviour of the new feature. Also, minor formatting changes have been made to improve code readability.