Refactor: Remove serviceClient as a class member in VirusScanValidator
serviceClient which is a CloseableHttpClient instance was removed as a class member and instantiated within the try-with-resources statement where it is used. This eliminates the need for managing its lifecycle manually, specifically the need for closing it during finalization. This change makes code cleaner and follows the principle of limiting the scope of variables as much as possible.