Add catalogIdentifier to Payload classes for improved context tracking
The code update introduces 'catalogIdentifier' as a parameter in various Payload classes under the Persistence module. This was initiated to provide an enhanced context when tracking actions performed on these Payloads. This also caters to the need of logging the correct catalog identifier for document actions. Alongside, minor change on correcting spelling mistake for parameter 'catalogIdentifier' in the method 'getGroupsForCatalog' in UserInfo.kt is included.