Fix conflict in select-type component (#5792)
The default "select" formly field was redefined by us, but when calling the print view dialog, the select field was reset to its origin, which lead to a lot of problems in the main form.

So the code was refactored to define a type with its own name to prevent this conflict.