Fix/Improve race condition during startup
The createDoc-plugin created in a rare case two create toolbar-buttons.
This happened because we visit the form page on startup where the plugins are registered. When accessing the form page the form plugins are initialized. There's also code for plugins that are registered late to activate them immediately. In that code it could happen, that the registerForm-function is called another time.