Improve dynamic behavior by manipulating form controls
When an action updates another field, which was created by ngx-formly then this can be done by updating the model and call an expensive updateModel-function.
The better way is to update the form control directly.
When updating a repeat-component however it's more complicated, since additional controls are created. Here it's better to use the updateModel-strategy.