More work on direct CSW publishing.
This change finishes publishing of single documents and
multiple documents (by scheduling) directly to a CSW
interface. It also introduces a transaction functionality
that makes it possible to remove orphaned datasets from
the CSW interface for scheduled indexing. The removal of
orphans is catalog specific and does not interfere with
datasets of other catalogs.

Still the removal of deleted or unpublished documents is
missing.