Fix delay of quartz job
The trigger has not been added to the database in time and will be picked up after the next check, which is after 30s. So we wait here a little bit and run the job in another thread. It's probably because of a transaction annotation that the trigger is not written immediately.