Fix error handling select-type in repeat-component (#8573)
When a repeat-component contains a select with a dynamic required field, which is calculated from "field.model..." then this field does not get destroyed correctly, and the change event handler is still there throwing an error in the following case: Dataset with one "Vektorformat" filled is loaded. Then load another dataset where "Vektorformat" is empty, but visible!

Using the formControl instead seems to fix this issue.